// resalt-frontend2/src/models/FilterOperand.ts
export enum FilterOperand {
	// eslint-disable-next-line no-unused-vars
	CONTAINS = 'c',
	// eslint-disable-next-line no-unused-vars
	NOT_CONTAINS = 'nc',
	// eslint-disable-next-line no-unused-vars
	EQUALS = 'e',
	// eslint-disable-next-line no-unused-vars
	NOT_EQUALS = 'ne',
	// eslint-disable-next-line no-unused-vars
	STARTS_WITH = 'sw',
	// eslint-disable-next-line no-unused-vars
	ENDS_WITH = 'ew',
	// eslint-disable-next-line no-unused-vars
	GREATER_THAN_OR_EQUAL = 'gte',
	// eslint-disable-next-line no-unused-vars
	LESS_THAN_OR_EQUAL = 'lte',
}